refactor: Inserter 패키지 이동
