+++ src/test/java/roomescape/dao/ThemeRepositoryTest.java
@@ -9,6 +9,7 @@ import roomescape.fixture.MemberFixture;
 import roomescape.fixture.ReservationTimeFixture;
 import roomescape.fixture.ThemeFixture;
 import roomescape.util.DatabaseCleaner;
+import roomescape.util.ReservationInserter;
 
 import java.util.List;
