+++ src/test/java/roomescape/util/ReservationInserter.java
@@ -1,7 +1,11 @@
-package roomescape.dao;
+package roomescape.util;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import roomescape.dao.MemberRepository;
+import roomescape.dao.ReservationRepository;
+import roomescape.dao.ReservationTimeRepository;
+import roomescape.dao.ThemeRepository;
 import roomescape.domain.reservation.*;
 import roomescape.domain.user.Member;
